+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/auto/LimelightPositionLoggingWithDrivetrain.kt
@@ -1,9 +1,9 @@
-package org.firstinspires.ftc.teamcode.opmodes
+package org.firstinspires.ftc.teamcode.opmodes.auto
 
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp
 import org.firstinspires.ftc.teamcode.internals.templates.BaseOpMode
-import org.firstinspires.ftc.teamcode.scripts.LimelightPositionLogging
-import org.firstinspires.ftc.teamcode.scripts.MecanumDriveScript
+import org.firstinspires.ftc.teamcode.scripts.auto.LimelightPositionLogging
+import org.firstinspires.ftc.teamcode.scripts.drive.MecanumDriveScript
 
 @TeleOp(name="Limelight Position Logging With Drivetrain", group = BaseOpMode.DEBUG_GROUP_NAME)
 class LimelightPositionLoggingWithDrivetrain: BaseOpMode() {
