+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/hardwaretest/ActiveIntake.kt
@@ -1,8 +1,8 @@
-package org.firstinspires.ftc.teamcode.opmodes
+package org.firstinspires.ftc.teamcode.opmodes.hardwaretest
 
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp
 import org.firstinspires.ftc.teamcode.internals.templates.BaseOpMode
-import org.firstinspires.ftc.teamcode.scripts.ContinuousServoScript
+import org.firstinspires.ftc.teamcode.scripts.hardware.ContinuousServoScript
 
 @TeleOp(name = "Active Intake", group = BaseOpMode.DEBUG_GROUP_NAME)
 class ActiveIntake: BaseOpMode() {
