+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/hardwaretest/LinearSlideTest.kt
@@ -1,8 +1,8 @@
-package org.firstinspires.ftc.teamcode.opmodes
+package org.firstinspires.ftc.teamcode.opmodes.hardwaretest
 
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp
 import org.firstinspires.ftc.teamcode.internals.templates.BaseOpMode
-import org.firstinspires.ftc.teamcode.scripts.ContinuousMotorScript
+import org.firstinspires.ftc.teamcode.scripts.hardware.ContinuousMotorScript
 
 
 @TeleOp(name = "Linear Slide Test", group = BaseOpMode.DEBUG_GROUP_NAME)
