+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/hardwaretest/SingleMotorTest.kt
@@ -1,9 +1,9 @@
-package org.firstinspires.ftc.teamcode.opmodes
+package org.firstinspires.ftc.teamcode.opmodes.hardwaretest
 
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp
 import org.firstinspires.ftc.teamcode.internals.hardware.HardwareManager
 import org.firstinspires.ftc.teamcode.internals.templates.BaseOpMode
-import org.firstinspires.ftc.teamcode.scripts.ContinuousMotorScript
+import org.firstinspires.ftc.teamcode.scripts.hardware.ContinuousMotorScript
 
 @TeleOp(name = "Single Motor Test", group = BaseOpMode.DEBUG_GROUP_NAME)
 class SingleMotorTest: BaseOpMode() {
