+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/robot/HangBot.kt
@@ -1,9 +1,9 @@
-package org.firstinspires.ftc.teamcode.opmodes
+package org.firstinspires.ftc.teamcode.opmodes.robot
 
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp
 import org.firstinspires.ftc.teamcode.internals.templates.BaseOpMode
-import org.firstinspires.ftc.teamcode.scripts.ContinuousMotorScript
-import org.firstinspires.ftc.teamcode.scripts.MecanumDriveScript
+import org.firstinspires.ftc.teamcode.scripts.hardware.ContinuousMotorScript
+import org.firstinspires.ftc.teamcode.scripts.drive.MecanumDriveScript
 
 
 @TeleOp(name = "HangBot", group = BaseOpMode.FULL_GROUP_NAME)
